requirements.txt
# Core Framework
gradio>=4.0.0
dash>=2.11.0
dash-bootstrap-components>=1.4.0
flask>=2.3.0
flask-cors

# Mapping and Visualization
folium>=0.14.0
plotly>=5.15.0

# Data Processing and Analysis
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Machine Learning and AI
tensorflow>=2.13.0
tensorflow-hub>=0.14.0
pillow>=10.0.0

# Database
psycopg2-binary>=2.9.0

# API Services
requests>=2.31.0
twilio>=8.10.0
libretranslatepy>=1.0.0

# Security and Authentication
python-jose>=3.3.0
bcrypt>=4.0.0

# Configuration and Environment
python-dotenv>=1.0.0

# Utilities
geopy>=2.3.0
aiohttp>=3.8.0
websockets>=11.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# Testing Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
responses>=0.23.0
coverage>=7.3.0 